requirements.txt
asgiref==3.8.1
boto3==1.35.26
botocore==1.35.26
certifi==2024.8.30
charset-normalizer==3.3.2
Django==3.2.25
django-cors-headers==4.4.0
djangorestframework==3.15.1
idna==3.8
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==1.26.20
razorpay==1.4.2